% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwt.cor.matrix.R
\name{biwt.cor.matrix}
\alias{biwt.cor.matrix}
\title{A function to compute the biweight mean vector and covariance matrix}
\usage{
biwt.cor.matrix(x, r, median = TRUE, full.init = TRUE)
}
\arguments{
\item{x}{an \code{n x g} matrix or data frame (\code{n} is the number of measurements, \code{g} is the number of observations (genes) )}

\item{r}{breakdown (\code{k/n} where \code{k} is the largest number of observations that can be replaced with arbitrarily large values while keeping the estimates bounded)}

\item{median}{a logical command to determine whether the initialization is done using the coordinate-wise median and MAD (TRUE) or using the minimum covariance determinant (MCD)  (FALSE).  Using the MCD is substantially slower.}

\item{full.init}{a logical command to determine whether the initialization is done for each pair separately (FALSE) or only one time at the beginning using the entire data matrix (TRUE).  Initializing for each pair separately is substantially slower.}
}
\value{
returns a list consisting of:
\item{biwt.corr.matrix}{a matrix of the biweight correlations.}
\item{biwt.NAid.matrix}{a matrix representing whether the biweight correlation was possible to compute (will be NA if too much data is missing or if the initializations are not accurate).  0 if computed accurately, 1 if NA.}
}
\description{
Compute a multivariate location and scale estimate based on Tukey's biweight weight function.
}
\examples{

samp.data <- MASS::mvrnorm(30,mu=c(0,0,0),Sigma=matrix(c(1,.75,-.75,.75,1,-.75,-.75,-.75,1),ncol=3))
r<-0.2 # breakdown

# To compute the 3 pairwise correlations in matrix form:

samp.bw.cor.mat <- biwt.cor.matrix(samp.data,r)
samp.bw.cor.mat
}
\references{
Hardin, J., Mitani, A., Hicks, L., VanKoten, B.; A Robust Measure of Correlation Between Two Genes on a Microarray, \emph{BMC Bioinformatics, 8}:220; 2007.
}
