% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwt.dist.matrix.R
\name{biwt.dist.matrix}
\alias{biwt.dist.matrix}
\title{A function to compute the biweight mean vector and covariance matrix}
\usage{
biwt.dist.matrix(x, r, median = T, full.init = T, absval = T)
}
\arguments{
\item{x}{an \code{n x g} matrix or data frame (\code{n} is the number of measurements, \code{g} is the number of observations (genes) )}

\item{r}{breakdown (\code{k/n} where \code{k} is the largest number of observations that can be replaced with arbitrarily large values while keeping the estimates bounded)}

\item{median}{a logical command to determine whether the initialization is done using the coordinate-wise median and MAD (TRUE) or using the minimum covariance determinant (MCD)  (FALSE).  Using the MCD is substantially slower.}

\item{full.init}{a logical command to determine whether the initialization is done for each pair separately (FALSE) or only one time at the beginning using the entire data matrix (TRUE).  Initializing for each pair separately is substantially slower.}

\item{absval}{a logical command to determine whether the distance should be measured as 1 minus the absolute value of the correlation (TRUE) or simply 1 minus the correlation (FALSE)}
}
\value{
a list consisting of
\item{biwt.dist.matrix} {a matrix of the biweight distances (default is 1 minus absolute value of the biweight correlation).}
\item{biwt.NAid.matrix} {a matrix representing whether the biweight correlation was possible to compute (will be NA if too much data is missing or if the initializations are not accurate).  0 if computed accurately, 1 if NA.}
}
\description{
Compute a multivariate location and scale estimate based on Tukey's biweight weight function.
}
\examples{
samp.data <- MASS::mvrnorm(30,mu=c(0,0,0),Sigma=matrix(c(1,.75,-.75,.75,1,-.75,-.75,-.75,1),ncol=3))
r<-0.2 # breakdown

# To compute the 3 pairwise distances in matrix form:
samp.bw.dist.mat <- biwt.dist.matrix(samp.data,r)
samp.bw.dist.mat

# To convert the distances into an element of class \code{dist}
as.dist(samp.bw.dist.mat$dist.mat)
}
\references{
Hardin, J., Mitani, A., Hicks, L., VanKoten, B.; A Robust Measure of Correlation Between Two Genes on a Microarray, \emph{BMC Bioinformatics, 8}:220; 2007.
}
