% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.int.p.R
\name{chi.int.p}
\alias{chi.int.p}
\title{Title}
\usage{
chi.int.p(p, a, c1)
}
\arguments{
\item{p}{a number}

\item{a}{a number}

\item{c1}{a number}
}
\value{
a number
}
\description{
Title
}
\examples{
chi.int.p(1,2,3)
}
