% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.int2.p.R
\name{chi.int2.p}
\alias{chi.int2.p}
\title{Internal function}
\usage{
chi.int2.p(p, a, c1)
}
\arguments{
\item{p}{a number}

\item{a}{a number}

\item{c1}{a number}
}
\value{
a number
}
\description{
Internal function
}
\examples{
chi.int2.p(2,3,4)
}
