% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erho.bw.p.R
\name{erho.bw.p}
\alias{erho.bw.p}
\title{Internal function}
\usage{
erho.bw.p(p, c1)
}
\arguments{
\item{p}{a number}

\item{c1}{a cutoff}
}
\value{
a number, The derivative of the expected value of rho
}
\description{
Internal function
}
\examples{
erho.bw.p(2,3)
}
