% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erho.bw.R
\name{erho.bw}
\alias{erho.bw}
\title{Internal function}
\usage{
erho.bw(p, c1)
}
\arguments{
\item{p}{a number}

\item{c1}{a cutoff}
}
\value{
a number, The expected value of rho
}
\description{
Internal function
}
\examples{
erho.bw(2,3)
}
