% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksolve.R
\name{ksolve}
\alias{ksolve}
\title{Internal function}
\usage{
ksolve(d, p, c1, b0)
}
\arguments{
\item{d}{a vector}

\item{p}{a number}

\item{c1}{a number}

\item{b0}{a number}
}
\value{
a number
}
\description{
Internal function
}
\examples{
ksolve(rnorm(20,.1,2),1,3,1)
}
