% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psibw.R
\name{psibw}
\alias{psibw}
\title{Internal function}
\usage{
psibw(x, c1)
}
\arguments{
\item{x}{a vector}

\item{c1}{a cutoff}
}
\value{
a vector
}
\description{
Internal function
}
\examples{
psibw(rnorm(10),3)
}
