% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejpt.bw.R
\name{rejpt.bw}
\alias{rejpt.bw}
\title{Internal function}
\usage{
rejpt.bw(p, r)
}
\arguments{
\item{p}{a number}

\item{r}{the breakdown}
}
\value{
the asymptotic rejection point
}
\description{
Internal function
}
\examples{
rejpt.bw(2,3)
}
