% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect2diss_hop.r
\name{vect2diss_hop}
\alias{vect2diss_hop}
\title{Internal function}
\usage{
vect2diss_hop(v)
}
\arguments{
\item{v}{a vector}
}
\value{
a vector
}
\description{
Internal function
}
\examples{
vect2diss_hop(rnorm(10))
}
