% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwt.est.R
\name{biwt.est}
\alias{biwt.est}
\title{A function to compute the biweight mean vector and covariance matrix}
\usage{
biwt.est(x, r, med.init)
}
\arguments{
\item{x}{an \code{n x 2} matrix or data frame (\code{n} is the number of observations)}

\item{r}{breakdown (\code{k/n} where \code{k} is the largest number of observations that can be replaced with arbitrarily large values while keeping the estimates bounded)}

\item{med.init}{a (robust) initial estimate of the center and shape of the data.  form is a list with components center and cov.}
}
\value{
A list with components
\item{biwt.mu}{the final estimate of location}
\item{biwt.sig}{the final estimate of scatter}
\item{biwt.NAid}{a logical of whether the given initialization was used (coded as 0) or whether a more precise initialization was used (namely, the pair by pair MCD, coded as 1).}
}
\description{
Compute a multivariate location and scale estimate based on Tukey's biweight weight function.
}
\examples{
samp.data <- MASS::mvrnorm(30,mu=c(0,0),Sigma=matrix(c(1,.75,.75,1),ncol=2))
r<-0.2 # breakdown

samp.mcd <- robustbase::covMcd(samp.data)
samp.bw <- biwt.est(samp.data,r,samp.mcd)

samp.bw.var1 <- samp.bw$biwt.sig[1,1]
samp.bw.var2 <- samp.bw$biwt.sig[2,2]
samp.bw.cov <- samp.bw$biwt.sig[1,2]

samp.bw.corr <- samp.bw.cov / sqrt(samp.bw.var1 * samp.bw.var2)

# or:

samp.bw.corr <- samp.bw$biwt.sig[1,2] / sqrt(samp.bw$biwt.sig[1,1]*samp.bw$biwt.sig[2,2])

##############
# to speed up the calculations, use the median/mad for the initialization:
##############

samp.init <- list()
samp.init$cov <- diag(apply(samp.data, 2, stats::mad, na.rm=TRUE))
samp.init$center <- apply(samp.data, 2, median, na.rm=TRUE)

samp.bw <- biwt.est(samp.data,r,samp.init)
samp.bw.corr <- samp.bw$biwt.sig[1,2] / sqrt(samp.bw$biwt.sig[1,1]*samp.bw$biwt.sig[2,2])
}
\references{
Hardin, J., Mitani, A., Hicks, L., VanKoten, B.; A Robust Measure of Correlation Between Two Genes on a Microarray, \emph{BMC Bioinformatics, 8}:220; 2007.
}
