% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwt.est.R
\name{biwt.est}
\alias{biwt.est}
\title{A function to compute the biweight mean vector and covariance matrix}
\usage{
biwt.est(x, r = 0.2, med.init = covMcd(x))
}
\arguments{
\item{x}{a \code{2 x n} matrix or data frame (\code{n} is the number of measurements)}

\item{r}{breakdown (\code{k/n} where \code{k} is the largest number of measurements that can be replaced with arbitrarily large values while keeping the estimates bounded). Default is \code{r = 0.2}.}

\item{med.init}{a (robust) initial estimate of the center and shape of the data.  The format is a list with components center and cov (as in the output of \code{covMcd()} from the rrcov package). Default is the minimum covariance determinant (MCD) on the data.}
}
\value{
A list with components
\item{biwt.mu}{the final estimate of center}
\item{biwt.sig}{the final estimate of shape}
}
\description{
Compute a multivariate location and scale estimate based on Tukey's biweight weight function.
}
\examples{
samp.data <- t(MASS::mvrnorm(30,mu=c(0,0),
               Sigma=matrix(c(1,.75,.75,1),ncol=2)))

samp.bw <- biwt.est(samp.data)
samp.bw

samp.bw.var1 <- samp.bw$biwt.sig[1,1]
samp.bw.var2 <- samp.bw$biwt.sig[2,2]
samp.bw.cov <- samp.bw$biwt.sig[1,2]

samp.bw.cor <- samp.bw.cov / sqrt(samp.bw.var1 * samp.bw.var2)
samp.bw.cor

# or:

samp.bw.cor <- samp.bw$biwt.sig[1,2] / sqrt(samp.bw$biwt.sig[1,1]*samp.bw$biwt.sig[2,2])
samp.bw.cor

##############
# to speed up the calculations, use the median/mad for the initialization:
##############

samp.init <- list()
samp.init$cov <- diag(apply(samp.data, 1, stats::mad, na.rm=TRUE))
samp.init$center <- apply(samp.data, 1, median, na.rm=TRUE)
samp.init

samp.bw <- biwt.est(samp.data, med.init = samp.init)
samp.bw.cor <- samp.bw$biwt.sig[1,2] / sqrt(samp.bw$biwt.sig[1,1]*samp.bw$biwt.sig[2,2])
samp.bw.cor
}
\references{
Hardin, J., Mitani, A., Hicks, L., VanKoten, B.; A Robust Measure of Correlation Between Two Genes on a Microarray, \emph{BMC Bioinformatics, 8}:220; 2007.
}
